requirements.txt
PyYAML
Flask==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
sentence-splitter==1.4
langdetect==1.0.7
summa==1.2.0
